+++ cli/tests/runner.rs
@@ -73,7 +73,6 @@ mod test_status_command;
 mod test_tag_command;
 mod test_templater;
 mod test_undo;
-mod test_unsquash_command;
 mod test_util_command;
 mod test_working_copy;
 mod test_workspaces;